% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{collections}
\alias{collections}
\alias{my_collections}
\alias{globus_ls}
\alias{mkdir}
\title{Globus collections, directories, and files}
\usage{
collections(filter_fulltext, filter_scope = "all", ..., all_fields = FALSE)

my_collections(filter_fulltext = "", ...)

globus_ls(
  .data,
  path = "",
  show_hidden = FALSE,
  filters = character(),
  all_fields = FALSE
)

mkdir(.data, path)
}
\description{
Globus collections, directories, and files
}
\examples{
##
## collections
##

collections <- collections("HuBMAP Public")  # '&' words or elements
collections

hubmap <-
    collections |>
    dplyr::filter(display_name == "HuBMAP Public")

my_collection <- my_collections()

##
## Directories and files
##

globus_ls(hubmap)    # same as `hubmap |> globus_ls()`

## hierarchical traversal
hubmap |>
    globus_ls("0008a49ac06f4afd886be81491a5a926/sprm_outputs")

## filter files with name ending in 'csv', and with size < 1000
hubmap |>
    globus_ls(
        "0008a49ac06f4afd886be81491a5a926/sprm_outputs",
        filters = "name:~*.csv/size:<1000"
    )

## filter files ending in 'json', or with size > 100000000
hubmap |>
    globus_ls(
        "0008a49ac06f4afd886be81491a5a926/sprm_outputs",
        filters = c("name:~*.json", "size:>100000000")
    )

\dontrun{
mkdir(my_collection, "tmp/HuBMAP/test")
}

}
